+++ lib/reaver/collection.rb
@@ -5,6 +5,7 @@ require 'time'
 require 'digest'
 
 module Reaver
+  # Loading collection from ~/.config/reaver/<filename.yml> and threat them
   class Collection
     attr_reader :tasks
 
@@ -33,15 +34,10 @@ module Reaver
       return unless @tasks || @tasks['things'].length.zero?
 
       @tasks['things'].each do |t|
-        if File.exist? t['name']
-          old_hash = Digest::MD5.file t['name']
-        else
-          @changed = true
-        end
-
+        hash_exist(t['name'])
         Reaver.download(t['url'], t['name'])
-        compare_hash(t['name'], old_hash) if old_hash
-
+        compare_to_old_hash(t['name']) if @old_hash
+        need_to_do_something_with(t) if @changed || !@old_hash
         metadata.info['changed'] = @changed
         @tasks['force_download'] = false
       end
@@ -49,9 +45,23 @@ module Reaver
 
     private
 
-    def compare_hash(filename, old_hash)
+    def hash_exist(file)
+      if File.exist? file
+        @old_hash = Digest::MD5.file file
+      else
+        @changed = true
+      end
+    end
+
+    def compare_to_old_hash(filename)
       hash = Digest::MD5.file filename
-      @changed = true if old_hash.hexdigest != hash.hexdigest
+      @changed = true if @old_hash.hexdigest != hash.hexdigest
+    end
+
+    def need_to_do_something_with(file)
+      dest = @tasks['all_to_dest_dir'] || file['dest_dir']
+      keep_name = @tasks['keep_name'] || false
+      Reaver::Walk.new(file['name'], dest, keep_name) if dest
     end
   end
 end